Refactor to IWinwingDevices
