+++ src/MobiFlightWwFcu/IWinwingDevice.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MobiFlightWwFcu
+{
+    internal interface IWinwingDevice
+    {        
+        void Connect();
+
+        void Shutdown();
+
+        List<string> GetLedNames();
+
+        List<string> GetDisplayNames();
+
+        void SetLed(string led, byte state);
+
+        void SetDisplay(string name, string value);
+    }
+}
